Revert "Refactor StateManager: Remove unused functions and centralize etcd operations in manager.rs"
