+++ src/player/statemanager/src/manager.rs
@@ -476,13 +476,7 @@ impl StateManagerManager {
             println!("  Associated Containers: {}", containers.len());
 
             // Get current model state from ETCD for comparison
-            let current_model_state = match Self::get_current_model_state(&model_name).await {
-                Ok(state) => state,
-                Err(e) => {
-                    eprintln!("  Failed to get current model state: {:?}", e);
-                    None
-                }
-            };
+            let current_model_state = StateMachine::get_current_model_state(&model_name).await;
 
             // Determine model state based on container states
             if let Some(new_model_state) = StateMachine::determine_model_state(&containers).await {
@@ -490,8 +484,9 @@ impl StateManagerManager {
 
                 // Check if model state actually changed
                 let state_changed = match current_model_state {
-                    Some(current_state) => current_state != new_model_state,
-                    None => true, // No previous state, so this is a change
+                    Ok(Some(current_state)) => current_state != new_model_state,
+                    Ok(None) => true, // No previous state, so this is a change
+                    Err(_) => true,   // Error getting state, treat as change
                 };
 
                 if state_changed {
@@ -500,7 +495,7 @@ impl StateManagerManager {
 
                     // Save the model state to etcd
                     if let Err(e) =
-                        Self::save_model_state_to_etcd(&model_name, new_model_state).await
+                        StateMachine::save_model_state_to_etcd(&model_name, new_model_state).await
                     {
                         eprintln!("  Failed to save model state to etcd: {:?}", e);
                     } else {
@@ -522,13 +517,28 @@ impl StateManagerManager {
             println!("  📈 Model states changed - processing package state updates");
 
             // Process package state changes based on model state changes
-            Self::process_package_state_changes().await;
+            StateMachine::process_package_state_changes().await;
         } else {
             println!("\n⏭️  No model state changes detected - skipping package state processing");
         }
 
         println!("  Status: Container list processing completed");
         println!("=====================================");
+    }
+
+    /// Creates a clone of self suitable for use in async tasks.
+    ///
+    /// This method provides a way to share the StateManagerManager instance
+    /// across multiple async tasks while maintaining proper ownership.
+    ///
+    /// # Returns
+    /// * `StateManagerManager` - Cloned instance for task use
+    fn clone_for_task(&self) -> StateManagerManager {
+        StateManagerManager {
+            state_machine: Arc::clone(&self.state_machine),
+            rx_container: Arc::clone(&self.rx_container),
+            rx_state_change: Arc::clone(&self.rx_state_change),
+        }
     }
 
     /// Runs the StateManagerManager's main event loop.
@@ -554,43 +564,10 @@ impl StateManagerManager {
         let arc_self = Arc::new(self);
         let grpc_manager = Arc::clone(&arc_self);
 
-        // Spawn the main processing task
+        // Spawn the main gRPC processing task
         let grpc_processor = tokio::spawn(async move {
-            // Main processing loop
-            loop {
-                tokio::select! {
-                    // Process container updates
-                    container_result = async {
-                        let mut rx = grpc_manager.rx_container.lock().await;
-                        rx.recv().await
-                    } => {
-                        match container_result {
-                            Some(container_list) => {
-                                grpc_manager.process_container_list(container_list).await;
-                            }
-                            None => {
-                                println!("Container channel closed, shutting down container processor");
-                                break;
-                            }
-                        }
-                    }
-
-                    // Process state change requests
-                    state_change_result = async {
-                        let mut rx = grpc_manager.rx_state_change.lock().await;
-                        rx.recv().await
-                    } => {
-                        match state_change_result {
-                            Some(state_change) => {
-                                grpc_manager.process_state_change(state_change).await;
-                            }
-                            None => {
-                                println!("State change channel closed, shutting down state processor");
-                                break;
-                            }
-                        }
-                    }
-                }
+            if let Err(e) = grpc_manager.process_grpc_requests().await {
+                eprintln!("Error in gRPC processor: {e:?}");
             }
         });
 
@@ -788,257 +765,6 @@ async fn execute_action(command: ActionCommand) {
         "  ✓ Action '{}' completed for: {}",
         command.action, command.resource_key
     );
-}
-
-// ========================================
-// ETCD STATE MANAGEMENT FUNCTIONS
-// ========================================
-// These functions handle all etcd operations for state persistence
-// and retrieval. Moved from state_machine.rs to centralize etcd
-// operations in the manager.
-
-impl StateManagerManager {
-    /// Retrieve all model states from ETCD using the /model/ prefix.
-    ///
-    /// # Returns
-    /// * `Result<Vec<(String, ModelState)>>` - Vector of (model_name, model_state) tuples
-    pub async fn get_all_model_states_from_etcd() -> Result<Vec<(String, ModelState)>> {
-        let prefix = "/model/";
-        println!("  Querying ETCD with prefix: {}", prefix);
-
-        let kvs = common::etcd::get_all_with_prefix(prefix).await?;
-        let mut model_states = Vec::new();
-
-        for kv in kvs {
-            // Extract model name from key: "/model/{name}/state" -> "{name}"
-            if let Some(model_name) = Self::extract_model_name_from_key(&kv.key) {
-                // Parse the state value
-                if let Ok(model_state) = Self::parse_model_state(&kv.value) {
-                    println!("    Found model: {} -> {:?}", model_name, model_state);
-                    model_states.push((model_name, model_state));
-                } else {
-                    eprintln!(
-                        "    Failed to parse model state for {}: {}",
-                        model_name, kv.value
-                    );
-                }
-            } else {
-                eprintln!("    Failed to extract model name from key: {}", kv.key);
-            }
-        }
-
-        println!("  Retrieved {} model states from ETCD", model_states.len());
-        Ok(model_states)
-    }
-
-    /// Extract model name from ETCD key format.
-    ///
-    /// # Arguments
-    /// * `key` - ETCD key in format "/model/{name}/state"
-    ///
-    /// # Returns
-    /// * `Option<String>` - Extracted model name or None if parsing fails
-    fn extract_model_name_from_key(key: &str) -> Option<String> {
-        if key.starts_with("/model/") && key.ends_with("/state") {
-            let name_part = &key[7..]; // Remove "/model/" prefix
-            if let Some(end) = name_part.rfind("/state") {
-                return Some(name_part[..end].to_string());
-            }
-        }
-        None
-    }
-
-    /// Parse model state string into ModelState enum.
-    ///
-    /// # Arguments
-    /// * `state_str` - String representation of model state
-    ///
-    /// # Returns
-    /// * `Result<ModelState>` - Parsed ModelState or error
-    fn parse_model_state(state_str: &str) -> Result<ModelState> {
-        match state_str {
-            "MODEL_STATE_CREATED" => Ok(ModelState::Created),
-            "MODEL_STATE_PAUSED" => Ok(ModelState::Paused),
-            "MODEL_STATE_EXITED" => Ok(ModelState::Exited),
-            "MODEL_STATE_DEAD" => Ok(ModelState::Dead),
-            "MODEL_STATE_RUNNING" => Ok(ModelState::Running),
-            _ => Err(format!("Unknown model state: {}", state_str).into()),
-        }
-    }
-
-    /// Get the current model state from ETCD for change detection.
-    ///
-    /// # Arguments
-    /// * `model_name` - Name/identifier of the model
-    ///
-    /// # Returns
-    /// * `Result<Option<ModelState>>` - Current model state if found, None if not found, or error
-    pub async fn get_current_model_state(model_name: &str) -> Result<Option<ModelState>> {
-        let key = format!("/model/{}/state", model_name);
-
-        match common::etcd::get(&key).await {
-            Ok(value) => {
-                // Parse the retrieved state value
-                match Self::parse_model_state(&value) {
-                    Ok(model_state) => Ok(Some(model_state)),
-                    Err(e) => {
-                        eprintln!(
-                            "  Failed to parse current model state for {}: {:?}",
-                            model_name, e
-                        );
-                        Err(e)
-                    }
-                }
-            }
-            Err(_) => {
-                // Key not found or other error - treat as no previous state
-                Ok(None)
-            }
-        }
-    }
-
-    /// Saves the model state to etcd using the specified key format.
-    ///
-    /// Implements the etcd storage pattern from the documentation:
-    /// Key format: /model/{model_name}/state
-    /// Value: ModelState as string representation
-    ///
-    /// # Arguments
-    /// * `model_name` - Name/identifier of the model
-    /// * `model_state` - The determined model state to save
-    ///
-    /// # Returns
-    /// * `Result<()>` - Success or error
-    pub async fn save_model_state_to_etcd(model_name: &str, model_state: ModelState) -> Result<()> {
-        let key = format!("/model/{}/state", model_name);
-        let value = model_state.as_str_name();
-
-        println!("    Saving to etcd - Key: {}, Value: {}", key, value);
-
-        if let Err(e) = common::etcd::put(&key, value).await {
-            eprintln!("    Failed to save model state to etcd: {:?}", e);
-            return Err(Box::new(e));
-        }
-
-        println!("    Successfully saved model state to etcd");
-        Ok(())
-    }
-
-    /// Saves package state to ETCD following the specified format.
-    ///
-    /// # Arguments
-    /// * `package_name` - Name/identifier of the package
-    /// * `package_state` - The determined package state to save
-    ///
-    /// # Returns
-    /// * `Result<()>` - Success or error
-    pub async fn save_package_state_to_etcd(
-        package_name: &str,
-        package_state: PackageState,
-    ) -> Result<()> {
-        let key = format!("/package/{}/state", package_name);
-        let value = package_state.as_str_name();
-
-        println!("    Saving to etcd - Key: {}, Value: {}", key, value);
-
-        if let Err(e) = common::etcd::put(&key, value).await {
-            eprintln!("    Failed to save package state to etcd: {:?}", e);
-            return Err(Box::new(e));
-        }
-
-        println!("    Successfully saved package state to etcd");
-        Ok(())
-    }
-
-    /// Process package state changes based on model states.
-    ///
-    /// This function implements the package state determination logic from the documentation,
-    /// retrieving all model states from ETCD and determining package states based on the
-    /// collective state of associated models:
-    /// - idle: No models (생성 시 기본 상태)
-    /// - paused: All models paused (모든 model이 paused 상태일 때)
-    /// - exited: All models exited (모든 model이 exited 상태일 때)
-    /// - degraded: Some models dead (일부 model이 dead 상태일 때)
-    /// - error: All models dead (모든 model이 dead 상태일 때)
-    /// - running: Default state (위 조건을 모두 만족하지 않을 때)
-    pub async fn process_package_state_changes() {
-        println!("--- Processing Package State Changes ---");
-
-        // Step 1: Get all model states from ETCD
-        let model_states = match Self::get_all_model_states_from_etcd().await {
-            Ok(states) => states,
-            Err(e) => {
-                eprintln!("  Failed to retrieve model states from ETCD: {:?}", e);
-                return;
-            }
-        };
-
-        if model_states.is_empty() {
-            println!("  No model states found in ETCD");
-            return;
-        }
-
-        println!("  Retrieved {} model states from ETCD", model_states.len());
-
-        // Step 2: Group models by package
-        // For now, we'll assume a simple naming convention: package = first part of model name before '-'
-        // In a real implementation, this would use proper metadata or annotations
-        let mut package_models: std::collections::HashMap<String, Vec<(String, ModelState)>> =
-            std::collections::HashMap::new();
-
-        for (model_name, model_state) in model_states {
-            // Extract package name from model name (simple convention)
-            let package_name = if let Some(pos) = model_name.find('-') {
-                model_name[..pos].to_string()
-            } else {
-                // If no '-' found, use the full model name as package name
-                model_name.clone()
-            };
-
-            package_models
-                .entry(package_name)
-                .or_insert_with(Vec::new)
-                .push((model_name, model_state));
-        }
-
-        println!(
-            "  Found {} packages with associated models",
-            package_models.len()
-        );
-
-        // Step 3: Process each package
-        for (package_name, models) in package_models {
-            println!("--- Processing Package: {} ---", package_name);
-            println!("  Associated Models: {}", models.len());
-
-            // Determine package state based on model states using StateMachine logic
-            if let Some(new_package_state) = StateMachine::determine_package_state(&models).await {
-                println!("  Determined Package State: {:?}", new_package_state);
-
-                // Save the package state to etcd
-                if let Err(e) =
-                    Self::save_package_state_to_etcd(&package_name, new_package_state).await
-                {
-                    eprintln!("  Failed to save package state to etcd: {:?}", e);
-                } else {
-                    println!("  Package state saved to etcd successfully");
-                }
-
-                // Check if ActionController reconcile is needed for error state
-                if matches!(new_package_state, PackageState::Error) {
-                    println!(
-                        "  📢 Package in ERROR state - would trigger ActionController reconcile"
-                    );
-                    // In a real implementation, this would send a gRPC request to ActionController
-                    // self.trigger_action_controller_reconcile(&package_name).await;
-                }
-            } else {
-                println!("  Could not determine package state from model states");
-            }
-        }
-
-        println!("Package state processing completed");
-    }
 }
 
 // ========================================
